services/delivery_worker/requirements.txt
asyncpg>=0.29.0
httpx>=0.27.0
pysnmp-lextudio>=6.0.0
